requirements.txt
python-telegram-bot==13.7
yt-dlp==2021.12.1
